docs(docker): Fix compose YAML schema type error
Needs to either be a list/array of strings containing the `=` character such as:

    - "DNS1=1.0.0.1"
    - "DNS2=1.1.1.1"

Or a mapping/dict/object such as in this change. I prefer this latter format because it
feels more consistent with the fact that environment variables can only have a single
value (as opposed to HTTP headers, for example) and the mapping YAML type
matches/enforces that restriction. It's also more readable IMO.

Resolves the following error:

    # docker compose run --rm --entrypoint bash dnsmasq
    services.dnsmasq.environment.[0]: unexpected type map[string]interface {}