docs(docker): Workaround common ignored query case
I figure answering queries from the hosts' LAN is a common use case and so this is worth
a FAQ answer. I don't think this can be added to the default configuration because the
user may use `network_mode: "host"` or other environments where the interface name can't
be assumed and I assume there are security risks for that default even if the interface
name could be assumed.